Spanner: Handle AbortedExceptions that occur before commit (#4181)
* spanner: Handle AbortedExceptions that occur before commit

The current TransactionRunner code checks for aborted transactions by
using the isAborted() function. This function returns the value of
"aborted" that is only set when processing resultSets. This does not
work for DML which only uses the resultSet stats and does not actually
call ResultSet.next().

This patch adds an additional check for the aborted exception type and
retries such transactions.

* throw before executeUpdate, verify flag is false after test

* Update google-cloud-clients/google-cloud-spanner/src/test/java/com/google/cloud/spanner/it/ITDMLTest.java

Co-Authored-By: nithinsujir <118742+nithinsujir@users.noreply.github.com>