test: fix flaky integration test
The ITClosedSessionTest could cause a flaky failure because it
deletes sessions on the server and then repeatedly executes queries
on the session until it returns a NOT_FOUND error. The actual deletion
could however happen exactly at the moment that the test query was
executed, which will cause Cloud Spanner to return a different error
message than the standard 'Session not found' error message.

Fixes #41