fix: stop sending RPCs on InstanceNotFound
The session pool should handle InstanceNotFound in the same way
as DatabaseNotFound errors, i.e. stop sending RPCs and also
consider an instance that has been re-created with the same name
as a different instance, and hence require the user to re-create
any database client before it can be used again.

Fixes #60