fix: stop preparing session on most errors
Most errors that occur during preparing a session for a read/write
transaction should be considered permanent, and should stop the
automatic preparing of sessions. Any subsequent call to get a read/write
session will cause an in-process BeginTransaction RPC to be executed. If
the problem has been fixed in the meantime, the RPC will succeed and the
automatic prepare of sessions will start again. Otherwise, the error is
propagated to the user.

Fixes #177