fix: use nanos to prevent truncation errors
SpannerRetryHelperTest.testExceptionWithRetryInfo could sometimes fail on some
environments. I have not been able to reproduce it locally, but one possible
reason for the failure could be truncation when converting from nanoseconds to
milliseconds. Now both the timeout and the verification use nanosecond precision.

This change also removes all uses of hamcres in this test file.

Fixes #200