fix: include an explicit version for javax-annotations-api
The Maven flatten plugin transforms the pom of the project from a set of imports
and transitive dependencies into a pom that contains all dependencies and versions
explicitly. This process did however miss the javax-annotations-api dependency that
is only included for JDK9 and higher.

Fixes https://github.com/GoogleCloudPlatform/java-docs-samples/issues/3139