test: get write sessions in one call
The test case SessionPoolTest.testMaintainerKeepsWriteProportion could fail
as it queried the session pool for the number of write sessions by doing two
calls; one for available write sessions and one for sessions being prepared.
If one of the sessions that was being prepared was returned to the pool between
the two calls, the test case would fail because it calculated the wrong number.

Fixes #264