test: check that page actually contains an element
The test wrongly assumed that if the previous page returned a nextPageToken
that the next call would always contain exactly one more result. It is however
possible that the result that had not yet been returned was deleted by another
process in the meantime, which would cause the next page to be empty.

Fixes #570