test: allow partial statements on the mock server (#568)
Allow a partial statement to be registered as the key for a statement result on the mock server. This makes it easier to mock results for statements that are generated by a framework with random or time based values, as only the first (static) part of the statement needs to be registered on the mock server.