test: test stream retries in combination with inlined begin
Adds an additional test case for inlined begin transaction when the first statement is a query.
Queries are executed using executeStreamingSql, and if the stream is broken, the original RPC
is retried with a resume code. That retry should not include a BeginTransaction option when the
transaction has already been initialized by the first RPC.