docs: documents resume on update database ddl
Instead of throwing an ALREADY_EXISTS error on the update database ddl
operation, we resume the original operation. This is necessary, because
the update database ddl is retryable. Because of this, we don't want to
confuse the caller by bubbling up an ALREADY_EXISTS error on a retry,
even though they used a non-existing operation id.