test: fixes DatabaseClientImpl stuck test
The DatabaseClientImpl#singleUseBoundAsync is sometimes getting stuck
during the execution. This is simple to reproduce by executing the test
class multiple times.
This is due to the static single threaded executor that is shared
among all the test cases.
In this fix, we create an executor in the setup method, avoiding the
sharing of the same amongt different tests.