test: reduce wait time for BackendExhaustedTest
The BackendExhaustedTest deliberately causes requests to get stuck on the mock
server. This causes the Spanner instance to refuse to shutdown nicely and causes
it to wait for 10 seconds. This can be circumvented by using shutdownNow() to
force the closure of all gRPC transport channels. This reduces the overall build
time with approx 10 seconds.