cleanup: remove unused transaction manager
The SingleUseTransaction in the Connection API previously used a
TransactionManager instead of a TransactionRunner, as this was the only way that it
was possible to implement a statement-specific timeout. This is now supported by the
client library, and the SingleUseTransaction had already been moved to using only a
TransactionRunner. There was however still an unused TransactionManager reference in
the code.

Fixes #852