fix: allows user-agent header with header provider (#871)
* fix: allows user-agent header with header provider

A bug was introduced, where if the caller tried to set a custom user
agent with a header provider an exception would be thrown (for duplicate
keys). Here, we merge the user agent set by the client along with the
one set by the library, instead of throwing such exception.

* test: adds test for default user agent

Tests if the default user agent is present in the user-agent header set
in the GapicSpannerRpc class.