test: removes flaky close result set test (#888)
This test was introduced in order to debug stuck transactions. The cause was unrelated though.
The test here is flaky, because there is a race condition between setting the stream field in the AbstractResultSet and checking it in the close method. Since the result set is not intended to be thread safe, we will not be synchronizing this variable.