feat: adds getValue to ResultSet (#1073)
* feat: adds getValue method to result set

Adds a generic getValue method to the result set. It can be used to
retrieve any type from the database.

* feat: removes custom int64 and float64 arrays

* feat: fixes clirr checks

Clirr does not support java 8 default implementations, so it thinks this
are breaking changes. These in fact are not breaking changes, since we
provide implementations in the interfaces.

* test: fixes GrpcResultSet serialize test

* feat: implements getValueInternal in result set

Adds a less intrusive implementation of getValueInternal for the
AbstractResultSet.

* fix: fixes struct value decoding

Uses the correct type for decoding structs from the result set

* feat: makes the structArray method public

Makes it public the method  to retrieve an array of structs from a Value

* test: adds tests for value literals

* feat: accepts null in arrays for getValue

* feat: allows nulls in values from result sets

* docs: fixes java doc for getValue