feat: exposes default leader in database, and leader options / replicas in instance config (#1283)
* feat: expose leader options in instance config

Exposes the list of leader options for an instance from an instance
config.

* feat: expose default leader in database

* test: add it test for instance leader options

* test: add it test for database default leader

* feat: create replica info / type model

* feat: expose replicas in instance config

* test: add it test for replicas in instance config

* test: skip leader tests for emulator

* test: skip emulator for instance admin test

Skips leader options / replicas checking related tests for the emulator.

* docs: add link for leader config accepted values

* docs: fix the link for replica infos.

* test: skip / removes leader tests

We decided not to have setting the default leader tests, because it
requires a multi-regional instance. The one we currently have for
running the tests is regional, so the cost of setup would not justify
the testing.
We also skip the get leader options test, since the feature is not
enabled in production yet.