fix: drop databases after sample tests
There was a bug on dropping databases / deleting backups after the
sample tests where we were trying to delete in the regional instance
first and if an exception was thrown, we would try to delete on the
second instance. This did not work, because the delete / drop operations
do not fail if the database / backup are not found in the instance, the
call simply returns.

In this PR we have applied both deletes one after the other to fix the
issue, instead of relying on the exceptions.