chore: configure backups test to run in nightly build (#1409)
* test: remove unused junit categories

* test: add slow test category

Add a new JUnit category: SlowTest. It will be used to execute tests
that are part of the nightly build.
Also refactors the categories hierarchy to be easier to setup.

* test: add maven profile to run slow tests

Configures a profile to execute only the slow tests. It will be invoked
in the nightly build.

* chore: configure nightly build for slow tests

Sets up the kokoro configuration for running slow tests nightly

* chore: ignore templated build.sh

* chore: fix linting violations

* test: reverts tracertest category removal

* chore: address PR comments