test: re-introduce parallel test execution (#1503)
* test: add timeout to it database admin test ops

Adds timeout to admin operations in the ITDatabaseAdminTest (database,
backup and restore ops)

* test: decrease it test timeout

We tried increasing the it test timeout to see if that could help with
the intermittent failures. It did not help, so we are rolling it back to
the original value.

* test: re-introduce parallel it test execution

We tried disabling parallel it test execution to see if it would help
with the intermittent failures we were seeing. This did not decrease the
failures, so we are rolling this back.