chore: support untyped values
Adds support for untyped values that can be used as statement
parameters. This is required for Spangres, as many native PG drivers
send (some) parameters without an explicit type, and expect the backend
to infer the type from the context. If we were to include a
(placeholder) type for such a parameter, the type inference on the
backend fails.