feat: support CREATE DATABASE in Connection API (#1845)
* feat: support CREATE DATABASE in Connection API

Adds support for the CREATE DATABASE statement in the Connection API.
The statement can only be used with a SingleUseTransaction (i.e. auto
commit mode). It is not supported in DDL batches. The database that is
created will have the same dialect as the current database that the user
is connected to.

Fixes #1884

* fix: add clirr ignore

* 🦉 Updates from OwlBot post-processor

See https://github.com/googleapis/repo-automation-bots/blob/main/packages/owl-bot/README.md

Co-authored-by: Owl Bot <gcf-owl-bot[bot]@users.noreply.github.com>