feat: support analyzeUpdate
Adds support for analyzeUpdate for DML statements, similarly to
analyzeQuery for DQL statements. Executing a DML statement in PLAN mode
only returns the query plan, but does not actually execute the
statement. Executing a DML statement in PROFILE mode executes the
statement and returns the query plan and execution statistics.

Fixes #1866