fix: PostgreSQL parser should not treat \ as an escape char
The PostgreSQL parser that removes comments and checks the type of
statement should not consider a backslash inside a quoted literal or
identifier as an escape character. Instead, only double occurrences of
the same quotes as the begin/end quote should be considered as an escape
inside a quoted literal or identifier.

Fixes #1920