feat: enable client to server compression (#2117)
Enable compression of network traffic from client to server.

See the below screenshots from Wireshark for tests that show that the compression works both ways. Note: The rate of compression will depend on the type of data that is being sent. The below example for client->server communication is for sending 1,000 very similar mutations to Spanner. That compresses very well with gzip. More random data will compress less.

### Server -> Client - No Compression
![server-_client identity](https://user-images.githubusercontent.com/1196707/196627944-4d6f34fa-78a9-4023-a848-51a0ffe5228e.png)

### Server -> Client - Gzip Compression
![server-_client gzip](https://user-images.githubusercontent.com/1196707/196627948-d2d574ed-3f69-4b3c-8613-c162e21e207d.png)

### Client -> Server - No Compression
![client-_server identity](https://user-images.githubusercontent.com/1196707/196627950-b01ecb88-f6e8-4fd9-b7f8-257ad4211ac9.png)

### Client -> Server - Gzip Compression
![client-_server gzip](https://user-images.githubusercontent.com/1196707/196627952-ecf8b4bd-cd44-4941-af55-6aecaf03f69e.png)