feat: support parallel queries in Connection API
Adds support for executing queries in parallel in the Connection API.

The Connection API already supported executing queries asynchronously, but this method
would always execute all queries sequentially, even if multiple queries were queued up
for async execution. This change adds a new method that users can use to specifically
request queries to be executed in parallel. The method can only be used for queries.
DML statements are not allowed, including DML statements with a returning clause.

The method has two limitations when executed in a read/write transaction:
1. If the first statement of a transaction is a parallel query, the query is still
   effectively executed as a sequential query. The reason for this is that the first
   query in a transaction is responsible for starting the transaction, and all
   following queries in the same transaction must wait for the first statement to
   return a transaction ID.
2. Parallel queries in read/write transactions disables automated retries of aborted
   transactions. Any aborted error that is returned by Cloud Spanner during a
   read/write transaction that uses parallel queries will be propagated to the
   application.