feat: savepoint
Adds support for savepoints to the Connection API. Savepoints use the internal retry
mechanism for read/write transactions to emulate actual savepoints. Rolling back to
a savepoint is guaranteed to always work, but resuming the transaction after rolling
back can fail if the underlying data has been modified, or if one or more of the
statements before the savepoint that was rolled back to returns non-deterministic
data.