test: fix mock server ARRAY<BYTES> parameter cast
The recent change that made decoding BYTES and ARRAY<BYTES> lazy did not
update all code paths in the mock Spanner server that is used for testing.
This meant that tests that tried to use ARRAY<BYTES> query parameters would
run into a ClassCastException. This problem only occurs in (internal) test
code and does not affect user code.