feat: capture stack trace for session checkout is now optional
The session pool by default captures the stack trace of the thread that
checks out a session of the pool, so this can be used in case the
session is leaked. This is done by creating an exception already at the
moment that the session is checked out. Some monitoring tools log the
creation of this exception, giving the impression that the application
is throwing a large number of errors, while the error is actually never
thrown.
This commit makes this capturing optional. The default is to capture the
call stack, but users can turn this off in the SessionPoolOptions.