perf: remove custom transport executor
Remove the custom executor provider that is set on the underlying generated
clients and instead use the internal core gRPC executor provider. The latter
is a shared executor provider for all gRPC channels that creates threads on
demand. This prevents the creation of unnecessary threads at startup, and can
reduce overall thread usage for applications that create multiple Spanner
instances during their lifetime.