Default Retention Period: Move null expire timestamp preconditions from Backup class to DatabaseAdminClient class
With the default retention period feature, null value for expire time
is a valid value. However, this is not be a valid value in the
updateBackup path. So, any pre-conditions on null values should be in
the API path and not in the object builder path. Thereby, moving
precondition checks from backup class to DatabaseAdminClient class.