perf: only capture the call stack if the call is actually async
Only capture the call stack of a caller of one of the Connection API
methods if the top-level call was an async call. For synchronous calls,
we do not need to additionally capture the call stack, as any exception
will include the current call stack of the synchronous method call.