docs: add sample for transaction timeouts
Our current samples show how to set timeout and retry settings for all
RPC invocations for a given RPC method, and how to set a timeout for a
single statement, but we did not have a sample for setting a timeout for
an entire transaction. This is something that customers have been asking
multiple times. This sample shows how a transaction timeout can be set
by using a gRPC context with a timeout.