feat: support PostgreSQL for autoConfigEmulator
The autoConfigEmulator=true flag in the Connection API can be used to
automatically connect to the emulator and automatically create the
instance and database that is being referenced. This makes running a
quick test on the emulator much easier, as all you need to do is to
configure the correct (JDBC) connection URL, and it will automatically
work. This mode would always create a GoogleSQL database. This change
adds support for creating a PostgreSQL database if the user specifically
sets the dialect in the connection string.