docs: improve timeout and retry sample
The timeout and retry sample used 60 seconds for both initial,
max, and total timeout. It also used a 1.0 multiplier for the
timeout value. This made it impossible to explain how to set
an increasing RPC timeout value. It also rendered adding the
DEADLINE_EXCEEDED error code as a retryable code superfluous.