chore: remove double client lib token header
Removes the double inclusion of the standard client lib token when no
custom client lib token has been set. Also reuses more of the standard
feature for setting the client lib token, instaed of appending the
standard token at the moment that we are creating the gRPC stub. This
should reduce the probability that anyone who might be using a custom
header or custom client in some way will get it wrong.