fix: respect SPANNER_EMULATOR_HOST env var when autoConfigEmulator=true
The Connection API would always use the default emulator host when
autoConfigEmulator=true was set in the connection string. This makes it
harder to override the host name for the emulator when you also want the
emulator to automatically create the instance and database that you
connect to, as the only way to set it is to add it to the connection
string.