test: remove partition token from tests
Remove the example partition token from the SQL test scripts, as the
tokens are not stable when the encoded classes are changed. E.g.
modifying the Value classes will cause these tests to fail with a
de-serialization error, as the serialized version in the token no longer
corresponds with the version in code.

Fixes the build error in #2736