perf: use virtual threads for Connection API
Use virtual threads for the Connection executor threads when
possible. Virtual threads were introduced in Java 21 and are
well suited for operations that are often blocked on I/O instead
of executing heavy computations. This is the typical thing that
the executor thread for connections do, as they are mainly
waiting for results to be returned from Cloud Spanner.