chore: fix session acquire timeout test
The test for session acquire timeout did not actually do what it
was supposed to do, as it did not check that a timeout was actually
registered. This again also caused it to busy-wait for 5 seconds
to times (MinSessions=0 and MinSessions=1). This fixes both the
actual test, and reduces the overall test time by about 10 seconds.