fix: use default query options with statement cache (#2860)
Statements that were executed using the Connection API and that were not
found in the statement cache (that is; it was the first time it was
executed), would not use the default query options that had been set for
the connection. This would mean that for example an optimizer version
that had been set for the connection would not be used the first time a
given query string would be executed using a connection. All subsequent
executions of the statement would use specified optimizer version.

This change ensures that both the first and all following executions of
the statement use the default query options that have been set.