chore: loosen restrictions setRetryAbortsInternally
Calling setRetryAbortsInternally(boolean) was unnecessarily strict, as there
is no reason why it should be disallowed to change the value while in
auto-commit mode and/or read-only mode. The fact that the flag does not have
any direct impact on those modes, does not mean that it should be disallowed
to set the value.

The fact that it was impossible to set the value while in auto-commit mode,
also means that it is impossible to disable it using a startup command for
a PostgreSQL connection to PGAdapter, as PGAdapter creates connections in
auto-commit mode.